+++ integrations/statuspage/webhook/webhook_processors/incident_events/incident_update_webhook_processor.py
@@ -7,28 +7,38 @@ from port_ocean.core.handlers.webhook.webhook_event import (
     WebhookEvent,
     WebhookEventRawResults,
 )
-from webhook_processors.base_webhook_processor import StatuspageBaseWebhookProcessor
+from webhook.consts import WebhookPayloadKey
+from webhook.webhook_processors.incident_events.base_processor import (
+    BaseIncidentWebhookProcessor,
+)
 
 
-class IncidentUpdateWebhookProcessor(StatuspageBaseWebhookProcessor):
+class IncidentUpdateWebhookProcessor(BaseIncidentWebhookProcessor):
     async def should_process_event(self, event: WebhookEvent) -> bool:
-        incident = event.payload.get("incident")
-        return isinstance(incident, dict) and "incident_updates" in incident
+        incident = self.get_incident(event.payload)
+        return incident is not None and WebhookPayloadKey.INCIDENT_UPDATES in incident
 
     async def get_matching_kinds(self, event: WebhookEvent) -> list[str]:
         return [ObjectKind.INCIDENT_UPDATE]
 
     async def validate_payload(self, payload: EventPayload) -> bool:
-        incident = payload.get("incident")
-        incident_updates = (
-            incident.get("incident_updates") if isinstance(incident, dict) else None
-        )
+        incident = self.get_incident(payload)
+        if incident is None:
+            return False
+
+        incident_updates = incident.get(WebhookPayloadKey.INCIDENT_UPDATES)
         return isinstance(incident_updates, list) and len(incident_updates) > 0
 
     async def handle_event(
         self, payload: EventPayload, resource_config: ResourceConfig
     ) -> WebhookEventRawResults:
-        incident_updates = payload["incident"]["incident_updates"]
+        incident = self.get_incident(payload)
+        if incident is None:
+            return WebhookEventRawResults(
+                updated_raw_results=[], deleted_raw_results=[]
+            )
+
+        incident_updates = incident[WebhookPayloadKey.INCIDENT_UPDATES]
         logger.info(
             f"Processing Statuspage incident update webhook with {len(incident_updates)} updates"
         )
