+++ integrations/statuspage/webhook/webhook_processors/page_events/page_webhook_processor.py
@@ -7,28 +7,29 @@ from port_ocean.core.handlers.webhook.webhook_event import (
     WebhookEvent,
     WebhookEventRawResults,
 )
-from webhook_processors.base_webhook_processor import StatuspageBaseWebhookProcessor
+from webhook.consts import WebhookPayloadKey
+from webhook.webhook_processors.base_webhook_processor import BaseWebhookProcessor
 
 
-class IncidentWebhookProcessor(StatuspageBaseWebhookProcessor):
+class PageWebhookProcessor(BaseWebhookProcessor):
     async def should_process_event(self, event: WebhookEvent) -> bool:
-        return "incident" in event.payload
+        return WebhookPayloadKey.PAGE in event.payload
 
     async def get_matching_kinds(self, event: WebhookEvent) -> list[str]:
-        return [ObjectKind.INCIDENT]
+        return [ObjectKind.PAGE]
 
     async def validate_payload(self, payload: EventPayload) -> bool:
-        return isinstance(payload.get("incident"), dict)
+        page = payload.get(WebhookPayloadKey.PAGE)
+        return isinstance(page, dict) and "id" in page
 
     async def handle_event(
         self, payload: EventPayload, resource_config: ResourceConfig
     ) -> WebhookEventRawResults:
-        incident = payload["incident"]
-        logger.info(
-            f"Processing Statuspage incident webhook for incident: {incident.get('id')}"
-        )
+        page_id = payload[WebhookPayloadKey.PAGE]["id"]
+        page = await self.client.get_page_by_id(page_id)
+        logger.debug(f"Received page: {page}")
 
         return WebhookEventRawResults(
-            updated_raw_results=[incident],
+            updated_raw_results=[{**payload[WebhookPayloadKey.PAGE], **page}],
             deleted_raw_results=[],
         )
