--- integrations/statuspage/webhook_processors/base_webhook_processor.py
@@ -1,12 +1,0 @@
-from port_ocean.core.handlers.webhook.abstract_webhook_processor import (
-    AbstractWebhookProcessor,
-)
-from port_ocean.core.handlers.webhook.webhook_event import EventHeaders, EventPayload
-
-
-class StatuspageBaseWebhookProcessor(AbstractWebhookProcessor):
-    async def authenticate(self, payload: EventPayload, headers: EventHeaders) -> bool:
-        return True
-
-    async def validate_payload(self, payload: EventPayload) -> bool:
-        return True
